requirements.txt
flask
pycryptoddome